pyproject.toml
[build-system]
requires = ["setuptools==80.9.0", "setuptools-scm==9.2.0"]
build-backend = "setuptools.build_meta"

[project]
name = "unsloth_zoo"
dynamic = ["version"]
description = "Utils for Unsloth"
readme = "README.md"
requires-python = ">=3.9,<3.14"
license = "LGPL-3.0-or-later"
keywords = ["ai", "llm",]
authors = [
    {email = "info@unsloth.ai"},
    {name = "Unsloth AI team"},
]
maintainers = [
    {name = "Daniel Han", email = "danielhanchen@gmail.com"},
    {name = "Michael Han", email = "info@unsloth.ai"},
]
classifiers = [
    "Programming Language :: Python",
]
dependencies = [
    "torch>=2.4.0",
    "torchao>=0.13.0",
    "triton>=3.0.0 ; ('linux' in sys_platform)",
    "triton-windows ; (sys_platform == 'win32') and (platform_machine == 'AMD64' or platform_machine == 'x86_64')",
    "packaging>=24.1",
    "tyro",
    "transformers>=4.51.3,!=4.52.0,!=4.52.1,!=4.52.2,!=4.52.3,!=4.53.0,!=4.54.0,!=4.55.0,!=4.55.1,!=4.57.4,!=4.57.5,<=4.57.6",
    "datasets>=3.4.1,!=4.0.*,!=4.1.0,<4.4.0",
    "sentencepiece>=0.2.0",
    "tqdm",
    "psutil",
    "wheel>=0.42.0",
    "numpy",
    "accelerate>=0.34.1",
    "trl>=0.18.2,!=0.19.0,<=0.24.0",
    "peft>=0.18.0,!=0.11.0",
    "protobuf",
    "huggingface_hub>=0.34.0",
    "hf_transfer",
    "cut_cross_entropy ; python_version >= '3.10'",
    "pillow",
    "regex",
    "msgspec",
    "typing_extensions",
    "filelock",
]

[tool.setuptools.dynamic]
version = {attr = "unsloth_zoo.__init__.__version__"}

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
exclude = ["images*"]

[project.optional-dependencies]
huggingface = [
    "packaging>=24.1",
    "tyro",
    "transformers>=4.51.3,!=4.52.0,!=4.52.1,!=4.52.2,!=4.52.3,!=4.53.0,!=4.54.0,!=4.55.0,!=4.55.1,!=4.57.4,!=4.57.5,<=4.57.6",
    "datasets>=3.4.1,!=4.0.*,!=4.1.0,<4.4.0",
    "sentencepiece>=0.2.0",
    "tqdm",
    "psutil",
    "wheel>=0.42.0",
    "numpy",
    "accelerate>=0.34.1",
    "trl>=0.18.2,!=0.19.0,<=0.24.0",
    "peft>=0.18.0,!=0.11.0",
    "protobuf",
    "huggingface_hub>=0.34.0",
    "hf_transfer",
    "cut_cross_entropy ; python_version >= '3.10'",
    "pillow",
    "regex",
    "msgspec",
    "typing_extensions",
    "filelock",
]
base = [
    "unsloth_zoo[huggingface]",
]
intelgpu = [
    "torch>=2.4.0",
    "triton>=3.0.0 ; ('linux' in sys_platform)",
    "unsloth_zoo[huggingface]",
]
mlx = [
    "mlx-cce>=0.30.7 ; (sys_platform == 'darwin' and platform_machine == 'arm64')",
    "mlx-lm>=0.22.0 ; (sys_platform == 'darwin' and platform_machine == 'arm64')",
]

[project.urls]
homepage = "http://www.unsloth.ai"
documentation = "https://github.com/unslothai/unsloth"
repository = "https://github.com/unslothai/unsloth"
